In UniformLinearMover, moved the state change for state variable 'velocity' to occur in doStartMove and doEndMove.  The previous behavior was to change velocity when giving the mover a command to move to a destination or go at a certain velocity (and even fire a property change), even when the motion state of the mover is paused or stopped.
This could break (flawed) client code that relies on getVelocity() returning a future velocity instead of the actual current velocity of the mover.  The test for Bug 730 is a demo of this. The test previously assumed the velocity changed upon calling moveTo, but now must first wait for doStartMove to occur before testing the results of motion calculations.

Simkit tests pass as before (with the noted modification)
DAFS tests pass as before