Completed the public interface extraction of EventList into a base interface: BasicEventList.
This will cause compile errors in some client code, but the change to client code is minimal, as described below.

Changed Schedule to be implented in terms of this interface rather than in terms of the concrete EventList class.  Of importance this means that Schedule.getEventList() now returns a BasicEventList, not an EventList.  Since BasicEventList is a complete replication of the public interface for EventList, this change should impact client code *only* by requiring the type of any assignments from getEventList() to change from EventList to BasicEventList. (Subclasses of EventList should not be affected).

Expressions like

EventList el = Schedule.getEventList(3);
must become

BasicEventList el = Schedule.getEventList(3) ;

Or else cast the return value of getEventList() to (EventList) when you know it is safe to do so.  For the moment this is almost always safe because Schedule only ever instantiates EventList.