No changes to the production code base.
Added the Actions.jar file to the repository (again), since it is desired to distribute Actions along with Simkit.  

Removed failing unit tests that tested a use-case that is not sensible (for random numbers with vector-valued seeds, the value returned by getSeed is meaningless, and will never match the value given to setSeed).

Removed junit.jar, since the presence of junit can generally be assumed for IDEs, or else is easily obtained and installed for command-line development environments.

(Kirk is committing these changes at Arnie's request)